+++ pages/releases.js
@@ -1,41 +1,42 @@
 import React from 'react'
+import styled from 'styled-components'
 import DocsLayout from '../components/DocsLayout'
 import { getReleases } from '../utils/githubApi'
 import md from '../components/md'
 import Anchor from '../components/Anchor'
 import Loading from '../components/Loading'
+import rem from '../utils/rem'
+import { getFormattedDate } from '../utils/dates'
 
 
-function getFormattedDate(releaseDate) {
-  const date = new Date(releaseDate)
-  const year = date.getFullYear()
-  let month = (1 + date.getMonth()).toString()
-  month = month.length > 1 ? month : '0' + month
-  let day = date.getDate().toString()
-  day = day.length > 1 ? day : '0' + day
+const ReleaseName = styled.span`
+  margin-top: ${rem(40)};
+  margin-bottom: ${rem(-20)};
+  display: block;
+`
 
-  return day + '/' + month + '/' + year
-}
-
-const Ecosystem = ({ releases, sidebarPages }) => (
+const Releases = ({ releases, sidebarPages }) => (
   <DocsLayout
     useDocsSidebarMenu={false}
     pages={sidebarPages}
     title="Releases"
     description="Styled Components Releases"
   >
+    {md`
+      Updating styled components is usually just an \`npm install\`, only major version might have breaking changes that will be noted in the [CHANGELOG](https://github.com/styled-components/styled-components/blob/master/CHANGELOG.md)
+    `}
     {releases ? releases.map(release =>
-      <div key={release.id}>
+      <section key={release.id}>
         <Anchor id={release.name.split('.').join("")}>
-          {release.name} <i>({getFormattedDate(release.created_at)})</i>
+          <ReleaseName>{release.name} <i>({getFormattedDate(release.created_at)})</i></ReleaseName>
         </Anchor>
         {md(release.body)}
-      </div>
+      </section>
     ) : <Loading />}
   </DocsLayout>
 )
 
-Ecosystem.getInitialProps = async () => {
+Releases.getInitialProps = async () => {
   const releases = await getReleases()
 
   return {
@@ -44,4 +45,4 @@ Ecosystem.getInitialProps = async () => {
   }
 }
 
-export default Ecosystem
+export default Releases
