+++ utils/dates.js
@@ -1,0 +1,10 @@
+export const getFormattedDate = unformattedDate => {
+  const date = new Date(unformattedDate)
+  const year = date.getFullYear()
+  let month = (1 + date.getMonth()).toString()
+  month = month.length > 1 ? month : '0' + month
+  let day = date.getDate().toString()
+  day = day.length > 1 ? day : '0' + day
+
+  return day + '/' + month + '/' + year
+}
